+++ src/firmware/application/io/buttons/Filter.h
@@ -29,7 +29,7 @@ namespace io
         public:
         ButtonsFilter() = default;
 
-        bool isFiltered(size_t index, uint8_t& numberOfReadings, uint32_t& states) override
+        bool isFiltered(size_t index, uint8_t& numberOfReadings, uint16_t& states) override
         {
             // this is a board-optimized debouncer
             // by the time processing of buttons takes place, more than 5ms has already passed
