+++ src/firmware/application/io/encoders/stub/Encoders.h
@@ -69,7 +69,7 @@ namespace io
             public:
             virtual ~HWA() = default;
 
-            virtual bool state(size_t index, uint8_t& numberOfReadings, uint32_t& states) = 0;
+            virtual bool state(size_t index, uint8_t& numberOfReadings, uint16_t& states) = 0;
         };
 
         class Filter
